backend/requirements.txt
django
djangorestframework
joblib
scikit-learn
imbalanced-learn
